requirements.txt
Pillow>=8.0.0
piexif>=1.1.3
pytest>=6.0.0